Refactor AuthController for better HTTP response handling
Add endpoint to retrieve subcategories by category ID

Added a new GET endpoint in the CategoryController to retrieve subcategories based on the provided category ID. If the category is not found, a NOT_FOUND status is returned with an appropriate error message. In case of any other exception, an INTERNAL_SERVER_ERROR status is returned along with a generic error message.